backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
pymysql==1.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
